Allow consumers of libiperf3 to get the JSON output for a just-completed test.
This changes the behavior of iperf_json_finish() so that it no longer
outputs JSON output, but saves the rendered output in a NUL-terminated
string buffer.  After calling iperf_run_server() or iperf_run_client(),
the client application should check iperf_get_test_json_output() to see
if it returns a non-NULL pointer.  If so, there is JSON data available
for it to print or otherwise consume.  The buffer is automatically
deallocated when the containing iperf_test structure is deallocated
with iperf_free_test().

Also adds a new API call iperf_get_test_outfile() to find the output
FILE* structure.

Modifies the iperf3 application to use the new API.  Users of iperf3
will not notice any functional change.

No effect in "normal" output mode (non-JSON).